requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
openai==1.3.7
anthropic==0.7.7
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-docx==0.8.11
reportlab==4.0.7
markdown==3.5.1